cloud-functions/market-close-feed/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
requests==2.*
feedparser==6.*
